Make update only replace and insert upsert, ignore conflicts and insert with fails allowed

+++ src/lib/doctrine-dbal-bulk/tests/Flow/Doctrine/Bulk/Tests/Context/DatabaseContext.php
@@ -43,6 +43,10 @@ final class DatabaseContext
                 continue;
             }
 
+            if (\str_contains($table->getName(), 'mysql')) {
+                continue;
+            }
+
             $this->connection->getSchemaManager()->dropTable($table->getName());
         }
     }
